(#114): tour 이름 가져오기
