+++ src/features/tour/hook/useStartTrip.ts
@@ -2,12 +2,19 @@ import { useNavigate } from 'react-router-dom';
 
 import type { GeoTripLocation } from '@/shared';
 
+interface StartTripButtonParams extends GeoTripLocation {
+  contentId: string;
+}
+
 export const useStartTrip = () => {
   const navigate = useNavigate();
 
-  const handleStartTrip = (destination: GeoTripLocation) => {
-    const { lng, lat } = destination;
-    navigate(`/map/destination?lnt=${lng}&lat=${lat}&vehicle=pedestrian`);
+  const handleStartTrip = ({ lng, lat, contentId }: StartTripButtonParams) => {
+    const encodedId = btoa(String(contentId));
+
+    navigate(
+      `/map/destination?lnt=${lng}&lat=${lat}&id=${encodedId}&vehicle=pedestrian`,
+    );
   };
 
   return { handleStartTrip };
